Optimal Drug Dosage Control Strategy Of Immune System Using Reinforcement Learning
The aim of the project is to find optimal drug dosage for a person based on their Immune system. In this article , an RL-based drug dosage control strategy is presented for immune systems to guarantee the number of tumor and immune cells reaches a specified level. This is designed using Java programming language.